[Osdev 702] Move countries to separate module (#36)
Move countries to separate module so that it becomes possible to use by
both Django and ContriCleaner


[OSDEV-702](https://opensupplyhub.atlassian.net/browse/OSDEV-702)

---------
