+++ src/django/api/serializers/facility/facility_claim_serializer.py
@@ -2,7 +2,7 @@ from rest_framework.serializers import (
   ModelSerializer,
   SerializerMethodField,
 )
-from ...countries import COUNTRY_NAMES
+from countries.lib.countries import COUNTRY_NAMES
 from ...models import FacilityClaim
