+++ src/django/api/serializers/facility/facility_list_item_serializer.py
@@ -2,7 +2,7 @@ from rest_framework.serializers import (
     ModelSerializer,
     SerializerMethodField,
 )
-from ...countries import COUNTRY_NAMES
+from countries.lib.countries import COUNTRY_NAMES
 from ...models import FacilityListItem
 from .facility_match_serializer import FacilityMatchSerializer
